requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
Flask==3.1.2
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
pypdfium2==5.4.0
python-dotenv==1.2.1
requests==2.32.5
schedule==1.2.2
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
